Выполнено разбиение на модули
