+++ index.html
@@ -234,6 +234,6 @@
       </section>
     </template>
 <script src="/js/function.js"></script>
-<script src="/js/main.js"></script>
+<script type="module" src="/js/main.js"></script>
 </body>
 </html>
